pyproject.toml
[project]
name = "unicompass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx",
    "playwright",
    "pydantic",
    "python-dotenv",
    "fastapi",
    "uvicorn[standard]",
    "supabase",
    "sse-starlette",
    "python-multipart",
    "rapidfuzz>=3.14.3",
    "pymupdf>=1.25.0",
]
